GreatestInThree

+++ Task 2 (by priyanka)/RupeeTODollar.java
@@ -3,7 +3,7 @@ public class RupeeTODollar {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.print("Enter Rupees : ");
-        float Rupees = sc.nextInt();
+        float Rupees = sc.nextFloat();
         
         System.out.println("Dollar = $" + Rupees/82.84);
     }
